requirements.txt
flask
jinja2
msgpack-python
numpy
pandas
pyyaml
six
